Changes everywhere. Biggest changes:
Implemented a kernel for managing tasks in a more realtime nature,
allowing you to eliminate any usage of delay() in code, giving MAJOR
performance benefits.Implemented a serial terminal shell. can be
used for administrating an operational system in realtime and more

Lots of restructuring done and much, much more to do.